+++ Pagination/BuilderRangePaginator.php
@@ -26,10 +26,14 @@ use Hector\Query\QueryBuilder;
  */
 class BuilderRangePaginator extends QueryRangePaginator
 {
+    use OptimizedFetchTrait;
+
     public function __construct(
         Builder $builder,
         bool $withTotal = true,
+        bool $optimized = false,
     ) {
+        $this->optimized = $optimized;
         parent::__construct($builder, $withTotal);
     }
 
@@ -38,6 +42,10 @@ class BuilderRangePaginator extends QueryRangePaginator
      */
     protected function fetchItems(QueryBuilder $builder): array
     {
+        if (true === $this->optimized) {
+            return $this->fetchItemsOptimized($builder);
+        }
+
         /** @var Builder $builder */
         return $builder->all()->getArrayCopy();
     }